Move Linux library support out of linux.go
To align with the Android.bp backend, move library support into its
own file, linux_cclibs.go.

Rename linux.go to linux_backend.go. This retains functions that are
used by multiple module types, and implementations of the simple
module types (aliases and resources).
