+++ Blueprints
@@ -70,7 +70,8 @@ bootstrap_go_package {
         "core/strip.go",
         "core/template.go",
         "core/toolchain.go",
-        "core/linux.go",
+        "core/linux_backend.go",
+        "core/linux_cclibs.go",
         "core/linux_generated.go",
         "core/linux_kernel_module.go",
     ],
